matrimony_backend/requirements.txt
asgiref==3.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==4.2.23
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.10
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==1.26.20
whitenoise==6.6.0
boto3==1.34.0
botocore==1.34.162
django-storages==1.14.2
s3transfer==0.9.0
jmespath==1.0.1
firebase-admin==6.4.0
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
httplib2==0.31.0
CacheControl==0.14.3
cachetools==6.2.0
msgpack==1.1.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.2.5
rsa==4.9.1
uritemplate==4.2.0
celery==5.3.6
redis==5.0.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
